tp employee

+++ crud_swing/src/main/java/org/example/dao/PersonDAO.java
@@ -1,12 +1,11 @@
 package org.example.dao;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.example.model.Person;
 import org.example.util.HibernateUtil;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
+import org.hibernate.query.Query;
 
 import javax.persistence.Entity;
 import java.util.List;
@@ -15,50 +14,57 @@ import java.util.List;
 @Entity
 public class PersonDAO {
 
-    public void savePerson(Person person) {
+    private Session session;
+
+    public PersonDAO() {
+        this.session = HibernateUtil.getSessionFactory().openSession();
+    }
+
+    public void savePerson(Object person) {
         Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+        try  {
             transaction = session.beginTransaction();
             session.save(person);
             transaction.commit();
+            System.out.println("Person added successfully");
         } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
 
     public void updatePerson(Person person) {
         Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+        try {
             transaction = session.beginTransaction();
             session.update(person);
             transaction.commit();
+            System.out.println("Person updated successfully");
         } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
 
-    public void deletePerson(Person person) {
+    public void deletePerson(Long id) {
         Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+        try  {
             transaction = session.beginTransaction();
-            session.delete(person);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
+            Person person = session.get(Person.class, id);
+            if (person != null) {
+                session.delete(person);
+                System.out.println("Person deleted successfully");
             }
+            transaction.commit();
+
+        } catch (Exception e) {
+            if (transaction != null) transaction.rollback();
             e.printStackTrace();
         }
     }
 
     public Person getPersonById(Long id) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+        try {
             return session.get(Person.class, id);
         } catch (Exception e) {
             e.printStackTrace();
@@ -67,11 +73,18 @@ public class PersonDAO {
     }
 
     public List<Person> getAllPersons() {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Person", Person.class).list();
+        try {
+            Query<Person> query = session.createQuery("FROM Person", Person.class);
+            return query.list();
         } catch (Exception e) {
             e.printStackTrace();
             return null;
+        }
+    }
+
+    public void close() {
+        if (session != null && session.isOpen()) {
+            session.close();
         }
     }
 }
